// web/src/generated/src/generated/accounts/supplyChainConfig.ts
export type SupplyChainConfig = {
  discriminator: ReadonlyUint8Array;
  admin: Address;
  fabricante: Address;
  auditorHw: Address;
  tecnicoSw: Address;
  escuela: Address;
  adminBump: number;
  adminPdaBump: number;
  nextTokenId: bigint;
  totalNetbooks: bigint;
  roleRequestsCount: bigint;
  fabricanteCount: bigint;
  auditorHwCount: bigint;
  tecnicoSwCount: bigint;
  escuelaCount: bigint;
};